Module Tezos_008_PtEdo2Zk_test_helpers.Context

type t =

| B of Block.t

| I of Incremental.t

val branch : t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Block_hash.t

val get_level : t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Raw_level.t, Tezos_error_monad.TzCore.error list) Stdlib.result

val get_endorsers : t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_services.Delegate.Endorsing_rights.t list, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val get_endorser : t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key_hash * int list, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val get_voting_power : t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key_hash -> int32 Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Error_monad.shell_tzresult Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val get_total_voting_power : t -> int32 Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Error_monad.shell_tzresult Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val get_bakers : t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key_hash list, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val get_seed_nonce_hash : t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Nonce_hash.t, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val get_seed : t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Seed.seed, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

Returns the seed of the cycle to which the block belongs to.

val get_constants : t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Constants.t, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

Returns all the constants of the protocol

val get_minimal_valid_time : t -> priority:int -> endorsing_power:int -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Time.t, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val get_baking_reward : t -> priority:int -> endorsing_power:int -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Tez.t, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val get_endorsing_reward : t -> priority:int -> endorsing_power:int -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Tez.t, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

module Vote : sig ... end

module Contract : sig ... end

module Delegate : sig ... end

val init : ?endorsers_per_block:int -> ?with_commitments:bool -> ?initial_balances:int64 list -> ?initial_endorsers:int -> ?min_proposal_quorum:int32 -> int -> (Block.t * Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t list, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

init n : returns an initial block with n initialized accounts and the associated implicit contracts