Module Context.Contract

val pp : Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Format.formatter -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> unit

val pkh : Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key_hash, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

type balance_kind =

| Main

| Deposit

| Fees

| Rewards

val balance : ?kind:balance_kind -> t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Tez.t, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

Returns the balance of a contract, by default the main balance. If the contract is implicit the frozen balances are available too: deposit, fees or rewards.

val counter : t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Z.t, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val manager : t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> (Account.t, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val is_manager_key_revealed : t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> (bool, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val delegate : t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key_hash, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t

val delegate_opt : t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key_hash option, Tezos_error_monad.TzCore.error list) Stdlib.result Tezos_protocol_008_PtEdo2Zk.Protocol.Environment.Lwt.t