Module Tezos_008_PtEdo2Zk_test_helpers.Op

val endorsement : ?delegate:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key_hash -> ?level:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Raw_level.t -> Context.t -> ?signing_context:Context.t -> unit -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Kind.endorsement Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val miss_signed_endorsement : ?level:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Raw_level.t -> Context.t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Kind.endorsement Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val transaction : ?counter:Z.t -> ?fee:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Tez.tez -> ?gas_limit:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> ?parameters:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Script.lazy_expr -> ?entrypoint:string -> Context.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Tez.t -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.packed, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val delegation : ?fee:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Tez.tez -> Context.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key_hash option -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.packed, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val revelation : ?fee:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Tez.tez -> Context.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.packed, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val origination : ?counter:Z.t -> ?delegate:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key_hash -> script:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Script.t -> ?preorigination:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.contract option -> ?public_key:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key -> ?credit:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Tez.tez -> ?fee:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Tez.tez -> ?gas_limit:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Gas.Arith.integral -> ?storage_limit:Z.t -> Context.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.contract -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.packed * Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.contract, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val originated_contract : Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.packed -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.contract

val double_endorsement : Context.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Kind.endorsement Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Kind.endorsement Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.packed

val double_baking : Context.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Block_header.block_header -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Block_header.block_header -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.packed

val activation : Context.t -> Tezos_crypto.Signature.Public_key_hash.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Blinded_public_key_hash.activation_code -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.packed, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val combine_operations : ?public_key:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.public_key -> ?counter:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.counter -> ?spurious_operation:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.packed_operation -> source:Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> Context.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.packed_operation list -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.packed_operation, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val seed_nonce_revelation : Context.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Raw_level.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Nonce.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.packed

Reveals a seed_nonce that was previously committed at a certain level

val proposals : Context.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> Tezos_crypto.Protocol_hash.t list -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.packed, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Propose a list of protocol hashes during the approval voting

val ballot : Context.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Contract.t -> Tezos_crypto.Protocol_hash.t -> Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Vote.ballot -> (Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Operation.packed, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Cast a vote yay, nay or pass

val dummy_script : Tezos_protocol_008_PtEdo2Zk.Protocol.Alpha_context.Script.t

val dummy_script_cost : Test_tez.Tez.t